Demo API/requirements.txt
fastapi
pydantic
supabase
python-dotenv
tensorflow==2.15.0
Pillow
numpy